+++ app/components/issues/IssueCard/styles.module.css
@@ -1,0 +1,161 @@
+.issueCard {
+  background: white;
+  border: 1px solid #ddd;
+  border-radius: 12px;
+  padding: 1.5rem;
+  transition: all 0.2s;
+  cursor: pointer;
+  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+}
+
+.issueCard:hover {
+  transform: translateY(-2px);
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.15);
+}
+
+.issueCard.resolved {
+  border-left: 4px solid #4caf50;
+}
+
+.issueCard.resolved:hover {
+  border-left-color: #388e3c;
+}
+
+.issueCard.unresolved {
+  border-left: 4px solid #f44336;
+}
+
+.issueCard.unresolved:hover {
+  border-left-color: #d32f2f;
+}
+
+.issueHeader {
+  display: flex;
+  align-items: center;
+  gap: 1rem;
+  margin-bottom: 1rem;
+}
+
+.issueStatus {
+  display: flex;
+  align-items: center;
+  gap: 0.5rem;
+}
+
+.issuePriority {
+  margin-left: auto;
+}
+
+.priority {
+  padding: 0.35rem 0.75rem;
+  border-radius: 16px;
+  font-size: 0.875rem;
+  font-weight: 500;
+}
+
+.priorityHigh {
+  background: #ffebee;
+  color: #c62828;
+}
+
+.priorityMedium {
+  background: #fff3e0;
+  color: #ef6c00;
+}
+
+.priorityLow {
+  background: #e8f5e9;
+  color: #2e7d32;
+}
+
+.status {
+  padding: 0.35rem 0.75rem;
+  border-radius: 16px;
+  font-size: 0.875rem;
+  font-weight: 500;
+}
+
+.statusResolved {
+  background: #e8f5e9;
+  color: #2e7d32;
+}
+
+.statusUnresolved {
+  background: #ffebee;
+  color: #c62828;
+}
+
+.issueTitle {
+  font-size: 1.25rem;
+  font-weight: 600;
+  color: #333;
+  margin-bottom: 0.75rem;
+  line-height: 1.4;
+}
+
+.issueContent {
+  color: #666;
+  font-size: 1rem;
+  line-height: 1.6;
+  margin-bottom: 1rem;
+}
+
+.issueFooter {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  margin-top: 1rem;
+  padding-top: 1rem;
+  border-top: 1px solid #eee;
+}
+
+.issueMeta {
+  display: flex;
+  gap: 1.5rem;
+}
+
+.issueDates {
+  color: #999;
+  font-size: 0.875rem;
+}
+
+.tags {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 0.5rem;
+}
+
+.tag {
+  display: inline-flex;
+  align-items: center;
+  gap: 0.5rem;
+  padding: 0.35rem 0.75rem;
+  background: #f5f5f5;
+  border-radius: 16px;
+  font-size: 0.875rem;
+  color: #666;
+  transition: all 0.2s;
+}
+
+.tag:hover {
+  background: #e0e0e0;
+}
+
+.tagIcon {
+  font-size: 1rem;
+}
+
+@media (max-width: 768px) {
+  .issueCard {
+    padding: 1rem;
+  }
+
+  .issueTitle {
+    font-size: 1.1rem;
+  }
+
+  .issueMeta {
+    flex-direction: column;
+    gap: 0.5rem;
+  }
+}
