+++ app/components/layout/Sidebar/styles.module.css
@@ -1,0 +1,69 @@
+.sidebar {
+  position: sticky;
+  top: 2rem;
+  height: fit-content;
+}
+
+.tagList {
+  background: white;
+  border: 1px solid #ddd;
+  border-radius: 12px;
+  padding: 1.5rem;
+  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+}
+
+.tagListTitle {
+  font-size: 1.25rem;
+  font-weight: 500;
+  color: #333;
+  margin-bottom: 1.5rem;
+  padding-bottom: 0.75rem;
+  border-bottom: 1px solid #ddd;
+}
+
+.tagListContent {
+  display: flex;
+  flex-direction: column;
+  gap: 0.5rem;
+}
+
+.tagItem {
+  display: flex;
+  align-items: center;
+  gap: 0.75rem;
+  padding: 0.75rem;
+  border-radius: 8px;
+  cursor: pointer;
+  transition: all 0.2s;
+  color: #666;
+  font-size: 1.1rem;
+}
+
+.tagItem:hover {
+  background: #f5f5f5;
+}
+
+.tagItem.active {
+  background: #e3f2fd;
+  color: #0070f3;
+}
+
+.tagCount {
+  margin-left: auto;
+  font-size: 1rem;
+  color: #999;
+  background: #f5f5f5;
+  padding: 0.35rem 0.75rem;
+  border-radius: 16px;
+}
+
+@media (max-width: 768px) {
+  .sidebar {
+    position: static;
+    margin-bottom: 2rem;
+  }
+
+  .tagList {
+    padding: 1rem;
+  }
+}
