+++ app/lib/constants.js
@@ -1,0 +1,15 @@
+// タグの列挙型
+export const TAGS = {
+  DEVELOPMENT: '開発',
+  DESIGN: 'デザイン',
+  DOCUMENTATION: 'ドキュメント',
+  INFRASTRUCTURE: 'インフラ',
+  OTHER: 'その他'
+};
+
+// 優先度の列挙型
+export const PRIORITY = {
+  HIGH: '高',
+  MEDIUM: '中',
+  LOW: '低'
+};
